backend logic for detective done

+++ server/Game/GameLogic.js
@@ -13,40 +13,40 @@ function startGameLogic(game) {
         
         //Logic for what to do every second
         switch(game.clock) {
-            case 0: //Everyone goes to sleep 
+            case 0://Everyone goes to sleep 
                 io.to(game.roomid).emit("System Message", {msg: "Everyone please go to sleep"});
                 endChat(game);
                 break;
-            case 2: //Prompt mafia to wake up
+            case 2://Prompt mafia to wake up
                 io.to(game.roomid).emit("System Message", {msg: "Mafia please wake up"});
                 break;
-            case 4: //Mafia turn 
+            case 4://Mafia turn 
                 mafiaTurn(game);
                 break;
-            case 64: //Prompt mafia to sleep
+            case 64://Prompt mafia to sleep
                 io.to(game.roomid).emit("System Message", {msg: "Mafia please go to sleep"});
                 break;
-            case 66: //Prompt detective to wake up
+            case 66://Prompt detective to wake up
                 io.to(game.roomid).emit("System Message", {msg: "Detective please wake up"});
                 break;
-            case 68: //Detective turn
+            case 68://Detective turn
                 detectiveTurn(game);
                 break;
-
+            case 100://Prompt detective to sleep 
+                io.to(game.roomid).emit("System Message", {msg: "Detective please go to sleep"});
+                break;
+            case 102://Prompt doctor to wake up
+                io.to(game.roomid).emit("System Message", {msg: "Doctor please wake up"});
+                break;
+            case 104://Doctor turn
+                doctorTurn(game);
+                break;
+            
         }
         game.clock++;
     }, 1000); 
 
-        //Start detective turn
-        // await sleep(2);
-        // io.to(game.roomid).emit("System Message", {msg: "Detective please wake up"});
-        //await detectiveTurn(game);
-        // io.to(game.roomid).emit("System Message", {msg: "Detective please go to sleep"});
         
-        //Start doctor turn
-        // await sleep(2);
-        // io.to(game.roomid).emit("System Message", {msg: "Doctor please wake up"});
-        //await doctorTurn(game);
         // io.to(game.roomid).emit("System Message", {msg: "Doctor please go to sleep"});
         
         // //Everyone wake up
@@ -57,6 +57,15 @@ function startGameLogic(game) {
 
     
     //send game over
+}
+
+/**
+ * Checks if game is over 
+ * @param {Object} game - Game Object
+ * @return True/false 
+ */
+function gameOver(game) {
+    // game.
 }
 
 /**
@@ -85,10 +94,14 @@ function mafiaTurn(game) {
         //Listen to when the mafia vote
         socket.on("Voted", ({uid})=> {
             try {
-                const victim = findPlayer(game, uid);
+                let victim;
+                let msg = player.getName+" voted to kill no one";
+                if (uid !== "No one") {
+                    victim = findPlayer(game, uid);
+                    msg = player.getName+" voted to kill "+victim.getName;
+                }
                 
                 //Display who each player voted to kill 
-                const msg = player.getName+" voted to kill "+victim.getName;
                 io.to(mafiaRoom).emit("System Message", {msg: msg});
                 
                 //Store the vote
@@ -98,13 +111,19 @@ function mafiaTurn(game) {
                     game.votes[uid]++;
                 }
 
-                //Speed up to the next phase if everyone already voted 
-                if (game.votes[uid] === game.mafiaCache.length) {
-                    //Everyone agreed to vote for no one 
-                    if (uid === "NO ONE") delete game.votes[uid]; 
-                    game.clock = 63;
-                } else if(Object.keys(game.votes).length === game.mafiaCache.length) {
+                //Speed up to the next phase if everyone already voted
+                if (game.votes[uid] === game.mafiaCache.length) {//Everyone agreed to vote the same player 
                     delete game.votes[uid];
+                    //Change the votes to normal array for easier access
+                    //Push the player object into the votes
+                    if (uid !== "No one") game.votes.push(victim);
+                    //Tell frontend to end the timer
+                    io.to(mafiaRoom).emit("End Timer");
+                    game.clock = 63;
+                } else if(Object.keys(game.votes).length === game.mafiaCache.length) {//Everyone voted different player
+                    delete game.votes[uid]; //no one will be killed
+                    //Tell frontend to end the timer
+                    io.to(mafiaRoom).emit("End Timer");
                     game.clock = 63;
                 }
 
@@ -122,13 +141,65 @@ function mafiaTurn(game) {
 }
 
 /**
- * Checks if game is over 
- * @param {Object} game - Game Object
- * @return True/false 
+ * Detective Operation
+ * @param {Object} game - Game Object 
  */
-function gameOver(game) {
-    // game.
+function detectiveTurn(game) {
+    const detective = game.detective;
+    //Detective is dead or disconnected already
+    if (game.detective === undefined) return;
+    detective.emit("System Message", {msg:"Please click on the player you would like to check"});
+    //Ask to vote on who to check, which makes the avator clickable
+    detective.emit("Please Vote", ({timer: 30}));
+    detective.on("Voted", ({uid})=> {
+        try {
+            if (uid !== "No one") {
+                const target = findPlayer(game, uid);
+                detective.emit("Flip", {role: target.getRole, uid: uid});
+            }
+        } catch(err) {//Forced disconnect client when they provide undefined uid as this shouldn't happen
+            console.log(err);
+            detective.emit("Forced Disconnect", {msg: "Unexpected error occurred"});
+            detective.disconnect();
+        }
+    });
 }
+
+/**
+ * Doctor Operation
+ * @param {Object} game - Game Object 
+ */
+function doctorTurn(game) {
+    const doctor = game.doctor[0];
+    const revive = game.doctor[1];
+    const posion = game.doctor[2];
+
+    //Doctor is dead or disconnected already
+    if (doctor === null) return;
+    
+    //Revive potion 
+    if (revive === 1) {
+        const victim = game.votes[0];
+        if (victim === undefined) {
+            doctor.emit("System Message", {msg:"No one died tonight\n"});
+        } else {
+            doctor.emit("System Message", {msg:victim.getName+" died tonight\n"});
+        }
+        // if () //Doctor can't save himself
+        doctor.emit("Save", ({timer: 30}));
+    } 
+
+    //Posion 
+    if (doctor[2] === 1 && revive === doctor[1]) {
+        doctor.emit("System Message", {msg:"Please click on the player you would like to posion"});
+        doctor.emit("Please Vote", ({timer: 30}));
+        doctor.on("Voted", ()=> {
+
+        });
+    }
+    doctor.emit("System Message", {msg:"You have used everything, you can skip to make the game faster"});
+}
+
 
 /**
  * Find player by uid
@@ -137,8 +208,8 @@ function gameOver(game) {
  * @returns Player Object
  */
 function findPlayer(game, uid) {
-    const victim = game.socketCache[uid];
-    if (victim === undefined) throw "Undefined player";
+    const victim = game.socketsCache[uid];
+    if (victim === undefined) throw new ReferenceError("Undefined player");
     return game.players[victim.id];
 }
